adjust creator mapping assumptions about fuller forms of names
In Archon, the fuller form of name field is also available for corporate and family names; therefore, this should be migrated for those name types as well. Since the fuller form of name field is not available for those agent types in ASpace, migrate the fuller form of name content as if it were a variant name instead.