add code to unpublish processed accessions if desired
In Archon, processed accessions are hidden from public users, even if they are set to be published. This code will override the publish setting from Archon with "false" if the accession is processed since ASpace does not use the same logic as Archon does for when to make accessions publically viewable. This feature can be turned on or off through a class variable in ASpaceMapper.