check creator date expression to see if begin and end dates can be found, adjust date type accordingly
in ASpace, the options for creator dates are "single" or "range". This change uses a new method, getNormalDate, to see if there are two years that are different from each other in the date expression. If there are, then the date is coded as a range and the begin and end dates are also recorded. If the begin and end dates from the normal date are the same, then it is coded as a single date with only the begin date logged.