feat: add methods to support parsing structured exted from string
This will support additional logic in the addResourceExtent method
to parse the alternative extent statments in cases where they were
used in Archon to document multiple extents using unstructured
language. Archon does not permit multiple extents, so the
alternative extent field was widely used to doument multiple extent
types.

The current behavior in addResourceExtent tread the alternative
extent closer to how it is defined in
https://saa-ts-dacs.github.io/dacs/06_part_I/03_chapter_02/05_extent.html
as an alternative way to describe the whole extent.