fix: portion should default to "part" if an alt extent exists
In practice, alt extent is used to descibe collections with
multiple extents. The presense of an alt extent should, in most
cases, be an indicator that the extents only describe part of the
collection. The exception is when the alternative extent is
obviously a subsidiary of the main extent, e.g., folders and items
where items count the number of things found in the folders.