fix: load existing repositories info from ASpace before copying repositories from Archon
Previously when doing repeat test imports the records would not end up in the correct repositories after the first time (at least when testing with multiple archon instances). The copy repository task would fail because the short code was already in use, and no information about the uri for the already in use repository would be available, so the records would end up in whatever repository was aspace id 2.