feat: always use the longest match for "contains" comparisons.
We want to avoid preferring one kind of match simply because it is
later in the list of existing extents. For, example when matching
microfilm reels against this list using a contains <|> contains
strategy, we don't want to end up with "reel" because it was the
last to match:

item
microfilm_reel
megabyte
reel

And for the inverse of this list, we don't want to stop mattching
when we get to reel because there might be a better match later on.