set the revision date by finding the end date of the normal date from the revision statement
This feature uses the getNormalDate method written for the collection content date parsing to determine what the last year referenced in the revision statement is. This year is placed in the date field and if no year is found then the revision is described as "undated" instead. Note that only four-digit years will be found with this method, to avoid misidentifying RS numbers as dates.